+++ src/main/java/tk/mybatis/mapper/MapperException.java
@@ -1,0 +1,23 @@
+package tk.mybatis.mapper;
+
+/**
+ * @author liuzh
+ */
+public class MapperException extends RuntimeException {
+    public MapperException() {
+        super();
+    }
+
+    public MapperException(String message) {
+        super(message);
+    }
+
+    public MapperException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public MapperException(Throwable cause) {
+        super(cause);
+    }
+
+}