+++ src/main/java/tk/mybatis/mapper/provider/ExampleProvider.java
@@ -49,6 +49,9 @@ public class ExampleProvider extends MapperTemplate {
     public String selectCountByExample(MappedStatement ms) {
         Class<?> entityClass = getEntityClass(ms);
         StringBuilder sql = new StringBuilder();
+        if(isCheckExampleEntityClass()){
+            sql.append(SqlHelper.exampleCheck(entityClass));
+        }
         sql.append(SqlHelper.selectCount(entityClass));
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
         sql.append(SqlHelper.exampleWhereClause());
@@ -65,6 +68,9 @@ public class ExampleProvider extends MapperTemplate {
     public String deleteByExample(MappedStatement ms) {
         Class<?> entityClass = getEntityClass(ms);
         StringBuilder sql = new StringBuilder();
+        if(isCheckExampleEntityClass()){
+            sql.append(SqlHelper.exampleCheck(entityClass));
+        }
         sql.append(SqlHelper.deleteFromTable(entityClass, tableName(entityClass)));
         sql.append(SqlHelper.exampleWhereClause());
         return sql.toString();
@@ -82,6 +88,9 @@ public class ExampleProvider extends MapperTemplate {
         //将返回值修改为实体类型
         setResultType(ms, entityClass);
         StringBuilder sql = new StringBuilder("SELECT ");
+        if(isCheckExampleEntityClass()){
+            sql.append(SqlHelper.exampleCheck(entityClass));
+        }
         sql.append("<if test=\"distinct\">distinct</if>");
         //支持查询指定列
         sql.append(SqlHelper.exampleSelectColumns(entityClass));
@@ -111,6 +120,9 @@ public class ExampleProvider extends MapperTemplate {
     public String updateByExampleSelective(MappedStatement ms) {
         Class<?> entityClass = getEntityClass(ms);
         StringBuilder sql = new StringBuilder();
+        if(isCheckExampleEntityClass()){
+            sql.append(SqlHelper.exampleCheck(entityClass));
+        }
         sql.append(SqlHelper.updateTable(entityClass, tableName(entityClass), "example"));
         sql.append(SqlHelper.updateSetColumns(entityClass, "record", true, isNotEmpty()));
         sql.append(SqlHelper.updateByExampleWhereClause());
@@ -126,6 +138,9 @@ public class ExampleProvider extends MapperTemplate {
     public String updateByExample(MappedStatement ms) {
         Class<?> entityClass = getEntityClass(ms);
         StringBuilder sql = new StringBuilder();
+        if(isCheckExampleEntityClass()){
+            sql.append(SqlHelper.exampleCheck(entityClass));
+        }
         sql.append(SqlHelper.updateTable(entityClass, tableName(entityClass), "example"));
         sql.append(SqlHelper.updateSetColumns(entityClass, "record", false, false));
         sql.append(SqlHelper.updateByExampleWhereClause());