+++ src/test/java/tk/mybatis/mapper/mapper/MybatisHelper.java
@@ -73,6 +73,8 @@ public class MybatisHelper {
                 // 3.3.1版本增加
                 config.setEnableMethodAnnotation(true);
                 config.setNotEmpty(true);
+                //校验Example中的类型是否一致
+                config.setCheckExampleEntityClass(true);
                 // 序列的获取规则,使用{num}格式化参数，默认值为{0}.nextval，针对Oracle
                 // 可选参数一共3个，对应0,1,2,分别为SequenceName，ColumnName, PropertyName
                 //config.setSeqFormat("NEXT VALUE FOR {0}");